Add files via upload
Class is now in its own file with function declarations in a function declaration file. Game now displays a game over screen when the user collides with a wall or the snake.
